feature/usuario
elimino delete.php y uso popup.php como mensaje antes de eliminar un usuario. Tambien dejo los nombres de archivos, clases y metodos acomodados
